refactor(web): fetch /browse panel data client-side instead of via RSC props
Blob, tree, and commit-diff panels previously fetched git data (file
source, folder contents, full commit diffs) in server components and
passed it as props across the client boundary. React flight-encodes
those props into the document at seconds of main-thread time for large
payloads (a large commit page measured 38MB / ~8s of encoding), which
starves the event loop and health probes under crawler load.

The panels now render a shell and fetch via existing JSON API routes
with react-query (per repo convention), plus a new /api/folder_contents
route. Documents drop to the ~70KB shell regardless of content size,
and JS-less crawlers never trigger the data fetch at all.
