+++ packages/web/src/app/(app)/browse/[...path]/components/treePreviewPanel/treePreviewPanel.tsx
@@ -1,10 +1,6 @@
-
-import { Separator } from "@/components/ui/separator";
 import { getRepoInfoByName } from "@/actions";
-import { PathHeader } from "@/app/(app)/components/pathHeader";
-import { getFolderContents } from "@/features/git/getFolderContentsApi";
 import { isServiceError } from "@/lib/utils";
-import { PureTreePreviewPanel } from "./pureTreePreviewPanel";
+import { TreePreviewPanelClient } from "./treePreviewPanelClient";
 
 interface TreePreviewPanelProps {
     path: string;
@@ -13,37 +9,23 @@ interface TreePreviewPanelProps {
 }
 
 export const TreePreviewPanel = async ({ path, repoName, revisionName }: TreePreviewPanelProps) => {
-    const [repoInfoResponse, folderContentsResponse] = await Promise.all([
-        getRepoInfoByName(repoName),
-        getFolderContents({
-            repoName,
-            revisionName: revisionName ?? 'HEAD',
-            path,
-        })
-    ]);
+    const repoInfoResponse = await getRepoInfoByName(repoName);
 
-    if (isServiceError(folderContentsResponse) || isServiceError(repoInfoResponse)) {
+    if (isServiceError(repoInfoResponse)) {
         return <div>Error loading tree preview</div>
     }
 
     return (
-        <>
-            <div className="flex flex-row py-1 px-2 items-center justify-between">
-                <PathHeader
-                    path={path}
-                    repo={{
-                        name: repoName,
-                        codeHostType: repoInfoResponse.codeHostType,
-                        displayName: repoInfoResponse.displayName,
-                        externalWebUrl: repoInfoResponse.externalWebUrl,
-                    }}
-                    pathType="tree"
-                    isFileIconVisible={false}
-                    revisionName={revisionName}
-                />
-            </div>
-            <Separator />
-            <PureTreePreviewPanel items={folderContentsResponse} />
-        </>
+        <TreePreviewPanelClient
+            path={path}
+            repoName={repoName}
+            revisionName={revisionName}
+            repo={{
+                name: repoInfoResponse.name,
+                codeHostType: repoInfoResponse.codeHostType,
+                displayName: repoInfoResponse.displayName,
+                externalWebUrl: repoInfoResponse.externalWebUrl,
+            }}
+        />
     )
 }
